% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\name{add_DECIMAL}
\alias{add_DECIMAL}
\title{Add decimal formatting}
\usage{
add_DECIMAL(x, k)
}
\arguments{
\item{x}{Numeric vector}

\item{k}{Number of decimal places}
}
\value{
Formatted character vector
}
\description{
Add decimal formatting
}
\keyword{internal}
