% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\docType{package}
\name{DiscreteMorseR-package}
\alias{DiscreteMorseR}
\alias{DiscreteMorseR-package}
\title{DiscreteMorseR: Discrete Morse Theory Analysis for 3D Meshes}
\description{
A package for computing discrete Morse vector fields, critical simplices,
and performing topological analysis of 3D mesh data using Discrete Morse
Theory.
}
\details{
The package provides functions to analyze 3D triangular meshes, compute
discrete Morse functions, identify critical simplices, and visualize
the Morse complex structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DijoG/DiscreteMorseR}
  \item Report bugs at \url{https://github.com/DijoG/DiscreteMorseR/issues}
}

}
\author{
Gergo Dioszegi
}
\keyword{internal}
