% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\name{get_lexIDLAB}
\alias{get_lexIDLAB}
\title{Compute lexicographically sorted simplex labels}
\usage{
get_lexIDLAB(df)
}
\arguments{
\item{df}{Data frame with label and idlabel columns}
}
\value{
Data table with lexicographically sorted labels
}
\description{
Compute lexicographically sorted simplex labels
}
\keyword{internal}
