% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\name{get_lowerSTAR_fast}
\alias{get_lowerSTAR_fast}
\title{Compute lower star filtration (SILENT VERSION for parallel)}
\usage{
get_lowerSTAR_fast(vertex, edge, face, dirout = NULL)
}
\arguments{
\item{vertex}{Vertex data}

\item{edge}{Edge data}

\item{face}{Face data}

\item{dirout}{Output directory (optional)}
}
\value{
List of lower star sets
}
\description{
Compute lower star filtration (SILENT VERSION for parallel)
}
\keyword{internal}
