% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\name{get_MESH}
\alias{get_MESH}
\title{Ultra-fast mesh preparation}
\usage{
get_MESH(vertices, faces)
}
\arguments{
\item{vertices}{Vertex data from alphahull}

\item{faces}{Face data from alphahull}
}
\value{
Mesh data expected by compute_morse_complex()
}
\description{
Ultra-fast mesh preparation
}
