% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\name{get_simplex_center}
\alias{get_simplex_center}
\title{Get center coordinates of a simplex}
\usage{
get_simplex_center(simplex, vertices)
}
\arguments{
\item{simplex}{Simplex identifier string}

\item{vertices}{Vertex coordinates}
}
\value{
Numeric vector of coordinates
}
\description{
Get center coordinates of a simplex
}
\keyword{internal}
