% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\name{get_simplexCENTER}
\alias{get_simplexCENTER}
\title{Fast simplex center calculation}
\usage{
get_simplexCENTER(simplex, vertices_matrix)
}
\arguments{
\item{simplex}{Simplex identifier string}

\item{vertices_matrix}{Vertex coordinates as numeric matrix (columns: X, Y, Z)}
}
\value{
Numeric vector of coordinates
}
\description{
Fast simplex center calculation
}
\keyword{internal}
