% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\name{plot_critical_points}
\alias{plot_critical_points}
\title{Plot critical points}
\usage{
plot_critical_points(critical, vertices, colors, size)
}
\arguments{
\item{critical}{Critical simplices}

\item{vertices}{Vertex coordinates}

\item{colors}{Colors for different simplex types}

\item{size}{Point size}
}
\description{
Plot critical points
}
\keyword{internal}
