% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\name{plot_gradient_field}
\alias{plot_gradient_field}
\title{Plot gradient vector field}
\usage{
plot_gradient_field(vector_field, vertices, color, length_scale, alpha)
}
\arguments{
\item{vector_field}{Morse vector field}

\item{vertices}{Vertex coordinates}

\item{color}{Arrow color}

\item{length_scale}{Scale factor for arrow length}

\item{alpha}{Arrow transparency}
}
\description{
Plot gradient vector field
}
\keyword{internal}
