% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\name{visualize_morse_2d_panel}
\alias{visualize_morse_2d_panel}
\title{Create multiple 2D projection plots}
\usage{
visualize_morse_2d_panel(
  morse_complex,
  point_alpha = 0.6,
  point_size = 1,
  max_points = 50000
)
}
\arguments{
\item{morse_complex}{Output from compute_morse_complex()}

\item{point_alpha}{Point transparency}

\item{point_size}{Point size}

\item{max_points}{Maximum points per plot}
}
\value{
List of ggplot2 objects
}
\description{
Create multiple 2D projection plots
}
