% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\name{visualize_morse_summary}
\alias{visualize_morse_summary}
\title{Create summary visualization of Morse complex}
\usage{
visualize_morse_summary(mesh, morse_complex, output_file = NULL)
}
\arguments{
\item{mesh}{Mesh object}

\item{morse_complex}{Output from compute_morse_complex()}

\item{output_file}{Optional file to save screenshot}
}
\value{
Multi-panel rgl visualization
}
\description{
Create summary visualization of Morse complex
}
