% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\name{compute_lowerSTAR_parallel}
\alias{compute_lowerSTAR_parallel}
\title{Compute lower star in parallel}
\usage{
compute_lowerSTAR_parallel(vertex, edge, face, output_dir = NULL, cores = 4)
}
\arguments{
\item{vertex}{Vertex data}

\item{edge}{Edge data}

\item{face}{Face data}

\item{output_dir}{Output directory}

\item{cores}{Number of cores (4-12)}
}
\value{
Combined lower star results
}
\description{
Compute lower star in parallel
}
