% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\name{get_lowerSTAR}
\alias{get_lowerSTAR}
\title{Compute lower star filtration for vertices}
\usage{
get_lowerSTAR(vertex, edge, face, dirout = NULL, cores = 1)
}
\arguments{
\item{vertex}{Vertex data}

\item{edge}{Edge data}

\item{face}{Face data}

\item{dirout}{Output directory (optional)}

\item{cores}{Number of cores (for consistency)}
}
\value{
List of lower star sets
}
\description{
Compute lower star filtration for vertices
}
\keyword{internal}
