% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteMorseR.R
\name{get_SIMPLICES}
\alias{get_SIMPLICES}
\title{Ultra-fast simplex extraction (C++ backend)}
\usage{
get_SIMPLICES(mesh, txt_dirout = "")
}
\arguments{
\item{mesh}{Mesh object from get_CCMESH()}

\item{txt_dirout}{Directory for output files (optional)}
}
\value{
List of processed simplices
}
\description{
Ultra-fast simplex extraction (C++ backend)
}
\keyword{internal}
